fix: define types for all entities, optimise structure
