+++ news-JS/package.json
@@ -26,7 +26,7 @@
         "copy-webpack-plugin": "^12.0.2",
         "css-loader": "^6.10.0",
         "dotenv-webpack": "^8.0.1",
-        "eslint": "^9.22.0",
+        "eslint": "^8.57.1",
         "eslint-config-prettier": "^9.1.0",
         "eslint-plugin-import": "^2.29.1",
         "eslint-plugin-prettier": "^5.1.3",
