+++ news-JS/src/components/view/appView.ts
@@ -1,28 +1,6 @@
 import News from './news/news';
 import Sources from './sources/sources';
-
-export interface NewsItem {
-    source: { name: string };
-    author?: string;
-    title: string;
-    description: string;
-    url: string;
-    urlToImage?: string;
-    publishedAt: string;
-}
-
-export type NewsResponse = {
-    articles: ReadonlyArray<NewsItem>;
-};
-
-export interface Source {
-    id: string;
-    name: string;
-}
-
-export type SourcesResponse = {
-    sources: ReadonlyArray<Source>;
-};
+import { NewsResponse, SourcesResponse } from '../../types';
 
 export default class AppView {
     private news: News;
